Doing screenshots differently, now copying all documentation figures
There are other things we could do with these figures, like display them
all, or display translated and untranslated side-by-side.  Also, now that
we have all the figures as well as a mapping from id to screenshot figure,
we can make app/module screenshots use something other than the first
screenshot in the document.  This is useful for things that use the User
Guide for documentation.