Adding git history and working 'svn info' stupidity
If you run 'svn info' on a file with an '@' in the name, SVN seems to think
that whatever is after the '@' is a revision.  Some time ago, I worked around
this by appending an '@' to all file names.  The problem is that this causes
'svn info' to hit the network, which makes get_file_revision slow.  Instead,
we'll only append the '@' if the file name contains an '@', which limites the
network penalty to a few select files.  Suboptimal, but better.