Allow po-file checking to be skipped based on timestamps
We didn't do this before, because it depends not only on the po file itself
having been changed, but also changes in the pot file, and the pot file can
depend on sources throughout the entire source tree.  What we're now doing
is storing the md5 of the pot file after the first empty line (to avoid the
headers, which contain POT-Creation-Date) and comparing both timestamps and
the pot md5.  Skipping po file processing when possible is a HUGE savings
in time.