Fixed odd graph-clipping bug in link box lists
If you have a list of link boxes with graphs, like those on a set page,
clicking the "more" link for ellipsized text causes the text cell to
grow horizontally, even though the text doesn't expand into the new
horizontal space.  The graph then gets pushed to the right and clipped.
Removing 'overflow: hidden' from the graph div prevents this.