Don't try to save a graph when outpath == None
This is an odd crasher that occurred because we implicitly assume that
every file under revision control has at least one revision associated
to it.  This seemed reasonable, except that it fails.  We get the list
of all revisions from a top-level log, and parse file names from that,
rather than trying to log each file.  For SVN, the top-level 'svn log'
doesn't give us revisions for svn:externals.

For now, we're just doing a None check to prevent a crash.  Handling
this kind of stuff will require some substantial architectural work,
and I'm not sure I want to deal with it.