Fancy no-offscreen sliding effect for info box collapse/expand
If you collapse or expand a box with a slide effect, and part of that box
is offscreen, things feel very unresponsive, because you wait for sliding
that you can't actually see.  This code does any offscreen height changes
immediately, then slides what you can actually see.  We should add this
to expanders in container boxes as well.