First pass at using a sidebar for set pages
This adds what was once in the Info tab into a persistent sidebar.
We now need to rework the tab interface of the main content so that
we don't have wrapped tabs.  We might load the main content using
AJAX so we don't have to redo the tab content each time.