Various improvements to graph comments
There's one tricky thing in this commit.  The div containing a graph was
previously overflow:hidden.  For some reason, this caused the graph arrows
to sit on top of comment divs when the arrows had an opacity set to <1.0.
This makes no sense to me, but doing overflow:normal on div.graph fixes
it.  Since we only need overflow set while sliding, we set it when the
slide starts, and reset it when it's finished.

This commit also fixes the issue that left the "next" arrow active on the
next-most graph.  I've also made graph comments a bit more descriptive.