Keep a cache of which things have been added to the queue already
When processing the history of a module, especially from the beginning
of time, we end up trying to queue the same people over and over again.
Queue.push will check to see if the object is already queued, but this
involves a database query.  We now cache locally everything we've added
in this process first, greatly reducing the number of database queries.