Adding a test plugin for the Evolution Quick Reference Card
This helps us test the assumptions built into the plugin infrastructure.
It will be helpful for designing the plugin system for both the docs and
i18n modules.  It will be especially helpful for i18n because it's not
based on POT files.  We'll see how deep that assumption runs.