Doing set tabs in applications, changes in app registration
I made register_application take an object rather than a class.  This
allows initialization functions to pass in additional information to
the applications.  We use this in set tabs to pass the count so we
don't have to compute it twice.

I've also added an additional initialize_application for application
modules.  This is called when the query string has the application
key set, allowing applications to register themselves only on those
requests when they're needed.  This is a performance issue, as some
tabs only get added based on database criteria.  We don't want to
query the database for useless stuff on every AJAX request.