[blip/web] Adding the basic web infrastructure in
With this change, the very basics of HTTP responses work.  That is
to say, I can generate a generic error page because nothing else
is implemented.  Next step, bring the real web logic over.

This also touched core and sweep because I changed the __init__ of
Response to require a Request to be passed in.