[configure.ac] Use `cmd` instead of $(cmd) and bump python req to 2.4
Ryan pointed out that certain older versions of bash were treating
the $( (stuff)) as $((stuff)).  I explicitly wrote it with the extra
space there so that it wouldn't be an arithmetic expression, but I
guess old bash is dumb.  Use backticks instead, and we can still use
$() on the inner command substitution.

Also, Python 2.4 is required for the classmethod and staticmethod
decorators.