New sweep error mechanism
Previously there was just a single error field on Branch, which just
doesn't work well for setting and clearing errors from different
sources. This commit adds an Error table that can record errors for
anything with an ident. You can easily catch, set, and clear errors
with 'with', and they're identified with the same sourcefunc logic
from Timestamp.